+++ apps/server/api/stripe/methods/controller.ts
@@ -1,5 +1,5 @@
 import { defineController } from './$relay';
-import { getPaymentMethods } from '$/domains/front/reservation/stripe';
+import { getPaymentMethods } from '$/domains/stripe';
 
 export default defineController(({ getPaymentMethods }), ({ getPaymentMethods }) => ({
   get: async({ user }) => getPaymentMethods(user.id),
