+++ apps/server/api/stripe/payment/capture/controller.ts
@@ -1,5 +1,5 @@
 import { defineController } from './$relay';
-import { capturePaymentIntents } from '$/domains/front/reservation/stripe';
+import { capturePaymentIntents } from '$/domains/stripe';
 
 export default defineController(({ capturePaymentIntents }), ({ capturePaymentIntents }) => ({
   post: async({ body }) => capturePaymentIntents(body.reservationId, body.isCancel),
