+++ apps/server/api/stripe/payment/controller.ts
@@ -1,5 +1,5 @@
 import { defineController } from './$relay';
-import { createPaymentIntents } from '$/domains/front/reservation/stripe';
+import { createPaymentIntents } from '$/domains/stripe';
 
 export default defineController(({ createPaymentIntents }), ({ createPaymentIntents }) => ({
   post: async({ body }) => createPaymentIntents(body.reservationId, body.paymentMethodsId),
